move specs store to server
